requirements.txt
actionable-recourse==1.0.1
asttokens==2.4.0
autograd==1.6.2
backcall==0.2.0
cffi==1.15.1
cmake==3.27.5
contourpy==1.1.1
cplex==22.1.1.0
cvxpy==1.3.2
cycler==0.11.0
decorator==5.1.1
ecos==2.0.12
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
fonttools==4.42.1
future==0.18.3
gurobipy==10.0.3
imageio==2.31.3
importlib-resources==6.0.1
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lime==0.2.0.1
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mip==1.12.0
Mosek==10.1.11
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
osqp==0.6.3
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pluggy==1.3.0
prettytable==3.9.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pyclustering==0.10.1.2
pycparser==2.21
pyDOE2==1.3.0
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qdldl==0.1.7.post0
scikit-image==0.21.0
scikit-learn==1.3.0
scikit-learn-extra==0.3.0
scipy==1.11.2
scs==3.2.3
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
sklearn==0.0.post9
stack-data==0.6.2
sympy==1.12
tdqm==0.0.1
threadpoolctl==3.2.0
tifffile==2023.9.18
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.6
zipp==3.17.0
